Refactor of mkchain
This is a pretty large update to mkchain to fix a number of issues and
make it more resilient. We have encountered many times where it fails to
correctly generate a valid chain and may only output a partial or empty
chain. After looking at a number of scenarios where this has happened I
figured it would be good to just refactor a lot of what it does and see
if we could improve the tool as a whole. One of the common scenarios
where an incomplete chain could be made is when an AIA URI is
encountered that leads to a PKCS#7 bundle rather than a single
certificate for an intermediate. This solves the issue by detecting the
format and extracting all of the certs if it's a bundle. Additionally
the code will have OpenSSL do the work to determine a correct/valid
chain instead of trying to build it ourselves... in theory this should
be much more robust.

The main reason there are so many changes is because I got carried away
with improvements and adding tests. Hopefully that means it will be
better and more resilient in the future.

Test plan:
  - Tests using rspec and linting using rubocop are both clean
  - See that generating chains from various leaf certificates works
    correctly:
    ```sh
    # Intermediate(s) only
    mkchain <certificate.crt>

    # Leaf and root included as well
    mkchain -lr <certificate.crt>

    # Use an older CA bundle to ensure the chain better supports older
    # clients
    mkchain -c 2023-05-30 <certificate.crt>
    ```

Refs: CSRE-3255
Co-authored-by: David Warkentin <dwarkentin@instructure.com>